Bamboo Plan
Adding validation for when it is the Final Stage.
